Rename util file.
